refactor: extract ProdDataListEmitter (Phase 4c step 3b)
Move the Data-list-shape product emission methods out of
ProductCaseSirTypeGenerator into a new ProdDataListEmitter:

  - genConstr (was genConstrDataConstr)
  - genSelect (was genSelectDataList)
  - genMatch  (was genMatchDataList)

ProductCaseSirTypeGenerator's genConstrLowered, genSelect, and
genMatch wrappers now delegate to ProdDataListEmitter; the type-keyed
typegen API surface is unchanged. ProdDispatch.genMatch's data-list
branch and the Lowering comment update accordingly.

selectMatchCase / retrieveConstrIndex / retrieveConstrDecl stay on
ProductCaseSirTypeGenerator during the migration and are referenced
cross-class from ProdDataListEmitter and ProdBuiltinPairEmitter.
