+++ scalus-core/shared/src/main/scala/scalus/compiler/sir/lowering/ProdDispatch.scala
@@ -634,8 +634,8 @@ object ProdDispatch {
       *
       *   - `(Prod|Sum)UplcConstr` → tag-ordered Case via `genMatchUplcConstr`.
       *   - `ProdDataList` / `ProdDataConstr` / `PackedDataList` / `PairIntDataList`
-      *     → `ProductCaseSirTypeGenerator.genMatchDataList` (Data-shape extraction
-      *     via `unConstrData` + `headList`/`tailList`, or just field projection).
+      *     → `ProdDataListEmitter.genMatch` (Data-shape extraction via
+      *     `unConstrData` + `headList`/`tailList`, or just field projection).
       *   - `ProdBuiltinPair(_, _)` → `ProdBuiltinPairEmitter.genMatch`
       *     (`Case` on Pair for V4+, `fstPair`/`sndPair` for V1-V3).
       *   - `OneElementWrapper(_)` → fall through to the per-type
@@ -660,8 +660,8 @@ object ProdDispatch {
                 typegens.SumUplcConstrSirTypeGenerator
                     .genMatchUplcConstr(matchData, loweredScrutinee, optTargetType)
             case ProdDataList | ProdDataConstr | PackedDataList | PairIntDataList =>
-                typegens.ProductCaseSirTypeGenerator
-                    .genMatchDataList(matchData, loweredScrutinee, optTargetType)
+                typegens.ProdDataListEmitter
+                    .genMatch(matchData, loweredScrutinee, optTargetType)
             case _: ProdBuiltinPair =>
                 typegens.ProdBuiltinPairEmitter
                     .genMatch(matchData, loweredScrutinee, optTargetType)
